+++ juntagrico/admins/forms/subscription_admin_form.py
@@ -14,12 +14,14 @@ class SubscriptionAdminForm(forms.ModelForm):
         model = Subscription
         fields = '__all__'
 
-    subscription_members = forms.ModelMultipleChoiceField(queryset=MemberDao.all_members(), required=False,
-                                                          widget=admin.widgets.FilteredSelectMultiple('Member', False))
+    subscription_members = forms.ModelMultipleChoiceField(
+        queryset=MemberDao.all_members(), required=False, widget=admin.widgets.FilteredSelectMultiple('Member', False),
+        label=Subscription.recipients_names.short_description)
 
     def __init__(self, *a, **k):
         forms.ModelForm.__init__(self, *a, **k)
-        self.fields['primary_member'].queryset = self.instance.recipients
+        if 'primary_member' in self.fields.keys():
+            self.fields['primary_member'].queryset = self.instance.recipients
         if self.instance.pk is None:
             self.fields['subscription_members'].queryset = MemberDao.members_for_create_subscription()
         elif self.instance.state == 'waiting':