+++ juntagrico/entity/member.py
@@ -34,14 +34,14 @@ class Member(JuntagricoBaseModel):
 
     iban = models.CharField('IBAN', max_length=100, blank=True, default='')
 
-    future_subscription = models.ForeignKey('Subscription', related_name='members_future', null=True, blank=True,
-                                            on_delete=models.SET_NULL)
-    subscription = models.ForeignKey('Subscription', related_name='members', null=True, blank=True,
-                                     on_delete=models.SET_NULL)
+    future_subscription = models.ForeignKey(
+        'Subscription', related_name='members_future', null=True, blank=True, on_delete=models.SET_NULL)
+    subscription = models.ForeignKey(
+        'Subscription', related_name='members', null=True, blank=True, on_delete=models.SET_NULL)
     old_subscriptions = models.ManyToManyField(
         'Subscription', related_name='members_old')
 
-    confirmed = models.BooleanField(_('bestätigt'), default=False)
+    confirmed = models.BooleanField(_('E-Mail-Adresse bestätigt'), default=False)
     reachable_by_email = models.BooleanField(
         _('Kontaktierbar von der Job Seite aus'), default=False)
 
@@ -50,8 +50,11 @@ class Member(JuntagricoBaseModel):
         _('Kündigüngsdatum'), null=True, blank=True)
     end_date = models.DateField(
         _('Enddatum'), null=True, blank=True)
-    inactive = models.BooleanField(_('inaktiv'), default=False)
-    notes = models.TextField(_('Notizen'), max_length=1000, blank=True)
+    inactive = models.BooleanField(_('inaktiv'), default=False,
+                                   help_text=_('Sperrt Login und entfernt von E-Mail-Listen'))
+    notes = models.TextField(
+        _('Notizen'), max_length=1000, blank=True,
+        help_text=_('Notizen für Administration. Nicht sichtbar für {}'.format(Config.vocabulary('member'))))
 
     @property
     def is_cooperation_member(self):