+++ test/test_depot_list_generation.py
@@ -1,6 +1,7 @@
 from io import StringIO
 
 from django.core.management import call_command
+from django.utils import timezone
 
 from test.util.test import JuntagricoTestCase